requirements.txt
python-binance
pandas
websocket_client
matplotlib
seaborn
pyyaml